Reduced Disk IO with noatime and nodiratime
noatime: Disables the updating of access time for both files and directories so that reading a file does not update their access time (atime).
nodiratime: Disables updating of access time when opening directories so that the access time is not modified when enumerating directories. This routine also checks that the object is a directory, which slows down the routine.